requirements.txt
Flask
Flask-SocketIO
eventlet
redis
rq
configobj
Pillow
imageio
pyserial
RPi.GPIO
ffmpeg
urllib3
pigpio